src/test/java/com/store/products1/domian/filters/FilterByProductsTest.java
package com.store.products1.domian.filters;

import com.store.products1.data.entities.ProductEntity;
import com.store.products1.domain.filters.FilterByProducts;
import com.store.products1.tools.MockProducts;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;

import java.util.List;

@RunWith(MockitoJUnitRunner.class)
public class FilterByProductsTest {

    private FilterByProducts filterByProducts;

    @Before
    public void setup(){

        filterByProducts = new FilterByProducts();
    }

    @Test
    public void filterByIdReturnOnlyOneResult(){

        Integer filter = 1;
        List<ProductEntity> productsEntity = MockProducts.getProductsEntity();

        List<ProductEntity> products = filterByProducts.filterById(productsEntity,filter);

        Assert.assertEquals(products.size(),1);

    }

    @Test
    public void filterByWordReturnListFiltered(){

        String filter = "xyz";
        List<ProductEntity> productsEntity = MockProducts.getProductsEntity();

        List<ProductEntity> products = filterByProducts.filterByWord(productsEntity,filter);

        Assert.assertEquals(products.size(),5);

    }

    @Test
    public void filterByEmptyWordReturnAnyResult(){

        String filter = "";
        List<ProductEntity> productsEntity = MockProducts.getProductsEntity();

        List<ProductEntity> products = filterByProducts.filterByWord(productsEntity,filter);

        Assert.assertEquals(products.size(),0);

    }

    @Test
    public void filterNullReturnAnyResult(){

        String filter = null;
        List<ProductEntity> productsEntity = MockProducts.getProductsEntity();

        List<ProductEntity> products = filterByProducts.filterByWord(productsEntity,filter);

        Assert.assertEquals(products.size(),0);

    }
}
